Adding consistent documentation to all of the packages * also refactored metric over into protocol package * also renamed types to poller since all the random bits got refactored out
