+++ check/check.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-// The check package contains an implementation file per check-type.
+// Package check contains an implementation file per check-type.
 //
 // A typical/starter check implementation file would contain something like:
 //
