+++ utils/timestamp.go
@@ -14,7 +14,6 @@
 // limitations under the License.
 //
 
-// Package utils Utility Functions
 package utils
 
 import "time"
